Add missing newlines
